Use __call__ over Class.execute() for commands
